ガチ完成

+++ van/before.cl
@@ -1,9 +1,9 @@
     #define SIZE
 #pragma OPENCL EXTENSION cl_khr_fp64:enable
 __kernel void
-calc(__global float *a,
-              __global float *b,
-              __global float *c)
+calc(__global double *a,
+              __global double *b,
+              __global double *c)
 {
 for(int i=0;i<SIZE;i++){;
         for(int j=0;j<SIZE;j++){
